OLOY-1027. Fixed points rounding.

See merge request open-loyalty/open-loyalty!64

+++ src/OpenLoyalty/Bundle/UserBundle/Service/EmailProvider.php
@@ -168,10 +168,10 @@ class EmailProvider
 
     /**
      * @param CustomerDetails $customer
-     * @param int             $availableAmount
-     * @param int             $pointsAdded
+     * @param float           $availableAmount
+     * @param float           $pointsAdded
      */
-    public function addPointsToCustomer(CustomerDetails $customer, int $availableAmount, int $pointsAdded)
+    public function addPointsToCustomer(CustomerDetails $customer, float $availableAmount, float $pointsAdded)
     {
         $subject = sprintf('%s - new points', $this->loyaltyProgramName);
